Fixed all

+++ Lesson_1/hepotenuse.cpp
@@ -5,12 +5,12 @@
 int main()
 {
 
-    float katet_1;
-    float katet_2;
+    double katet_1;
+    double katet_2;
 
-    std:: cin >> katet_1 >> katet_2;
-    float hepotenuse = sqrt(katet_1 * katet_1 + katet_2 * katet_2);
-    std:: cout << hepotenuse << std::endl;
+    std::cin >> katet_1 >> katet_2;
+    double hepotenuse = sqrt(katet_1 * katet_1 + katet_2 * katet_2);
+    std::cout << hepotenuse << std::endl;
 
     return 0;
 }
